+++ test/performance/fcollect_bw.cpp
@@ -4,17 +4,17 @@
 
 #define BW_TEST_HEADER
 #define BW_TEST_FUNCTION                                                                           \
-    for (int i = 0; i < iterations; i += 1) {                                                      \
+    for (size_t i = 0; i < iterations; i += 1) {                                                   \
         ishmem_long_fcollect(ISHMEM_TEAM_WORLD, (long *) dest, (long *) src, nelems);              \
     }
 
 #define BW_TEST_FUNCTION_ON_QUEUE                                                                  \
-    for (int i = 0; i < iterations; i += 1) {                                                      \
+    for (size_t i = 0; i < iterations; i += 1) {                                                   \
         ishmemx_long_fcollect_on_queue((long *) dest, (long *) src, nelems, test_return, q);       \
     }
 
 #define BW_TEST_FUNCTION_WORK_GROUP                                                                \
-    for (int i = 0; i < iterations; i += 1) {                                                      \
+    for (size_t i = 0; i < iterations; i += 1) {                                                   \
         ishmemx_long_fcollect_work_group(ISHMEM_TEAM_WORLD, (long *) dest, (long *) src, nelems,   \
                                          grp);                                                     \
     }
