+++ test/performance/get_nbi_bw.cpp
@@ -4,13 +4,13 @@
 
 #define BW_TEST_HEADER int pe = n_pes - 1;
 #define BW_TEST_FUNCTION                                                                           \
-    for (int i = 0; i < iterations; i += 1) {                                                      \
+    for (size_t i = 0; i < iterations; i += 1) {                                                   \
         ishmem_long_get_nbi((long *) dest, (long *) src, nelems, pe);                              \
     }                                                                                              \
     ishmem_quiet();
 
 #define BW_TEST_FUNCTION_WORK_GROUP                                                                \
-    for (int i = 0; i < iterations; i += 1) {                                                      \
+    for (size_t i = 0; i < iterations; i += 1) {                                                   \
         ishmemx_long_get_nbi_work_group((long *) dest, (long *) src, nelems, pe, grp);             \
     }                                                                                              \
     ishmemx_quiet_work_group(grp);
